// src/app/services/models/position-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DepartementShortDto } from '../models/departement-short-dto';
export interface PositionDto {
  active?: boolean;
  department?: DepartementShortDto;
  description?: string;
  grade?: string;
  id?: number;
  maxSalary?: number;
  minSalary?: number;
  requirements?: string;
  responsibilities?: string;
  title?: string;
}